remove border spec types and update needed overloads
