+++ clique-core/src/main/java/io/github/kusoroadeolu/clique/CliqueComponents.java
@@ -2,15 +2,13 @@ package io.github.kusoroadeolu.clique;
 
 import io.github.kusoroadeolu.clique.boxes.Box;
 import io.github.kusoroadeolu.clique.boxes.BoxType;
-import io.github.kusoroadeolu.clique.boxes.DefaultBox;
 import io.github.kusoroadeolu.clique.config.*;
-import io.github.kusoroadeolu.clique.frame.DefaultFrame;
 import io.github.kusoroadeolu.clique.frame.Frame;
-import io.github.kusoroadeolu.clique.indent.DefaultIndenter;
 import io.github.kusoroadeolu.clique.indent.Indenter;
 import io.github.kusoroadeolu.clique.progressbar.IterableProgressBar;
 import io.github.kusoroadeolu.clique.progressbar.ProgressBar;
 import io.github.kusoroadeolu.clique.progressbar.ProgressBarPreset;
+import io.github.kusoroadeolu.clique.spi.AnsiCode;
 import io.github.kusoroadeolu.clique.tables.AbstractTable.TableHeaderBuilder;
 import io.github.kusoroadeolu.clique.tables.TableFactory;
 import io.github.kusoroadeolu.clique.tables.TableType;
@@ -45,51 +43,63 @@ final class CliqueComponents {
         return TableFactory.getTableBuilder(type);
     }
 
-    public static TableHeaderBuilder table(BorderSpec style) {
-        var borderStyle = BorderStyle.fromSpec(style);
-        return table(TableConfiguration.fromBorderStyle(borderStyle));
+    public static TableHeaderBuilder table(String borderColor) {
+        return table(TableConfiguration.builder().borderColor(borderColor).build());
     }
 
-    public static TableHeaderBuilder table(TableType type, BorderSpec style) {
-        var borderStyle = BorderStyle.fromSpec(style);
-        return table(type, TableConfiguration.fromBorderStyle(borderStyle));
+    public static TableHeaderBuilder table(TableType type, String borderColor) {
+        return table(type, TableConfiguration.builder().borderColor(borderColor).build());
+    }
+
+    public static TableHeaderBuilder table(AnsiCode... borderColor) {
+        return table(TableConfiguration.builder().borderColor(borderColor).build());
+    }
+
+    public static TableHeaderBuilder table(TableType type, AnsiCode... borderColor) {
+        return table(type, TableConfiguration.builder().borderColor(borderColor).build());
     }
 
     // BOX
     public static Box box() {
-        return new DefaultBox();
+        return new Box();
     }
 
     public static Box box(BoxConfiguration configuration) {
-        return new DefaultBox(configuration);
+        return new Box(configuration);
     }
 
     public static Box box(BoxType type, BoxConfiguration configuration) {
-        return new DefaultBox(type, configuration);
+        return new Box(type, configuration);
     }
 
     public static Box box(BoxType type) {
-        return new DefaultBox(type);
+        return new Box(type);
     }
 
-    public static Box box(BorderSpec style) {
-        var borderStyle = BorderStyle.fromSpec(style);
-        return new DefaultBox(BoxType.ROUNDED, BoxConfiguration.fromBorderStyle(borderStyle));
+    public static Box box(AnsiCode... borderColor) {
+        return new Box(BoxType.ROUNDED, BoxConfiguration.builder().borderColor(borderColor).build());
     }
 
-    public static Box box(BoxType type, BorderSpec style) {
-        var borderStyle = BorderStyle.fromSpec(style);
-        return new DefaultBox(type, BoxConfiguration.fromBorderStyle(borderStyle));
+    public static Box box(BoxType type, AnsiCode... borderColor) {
+        return new Box(type, BoxConfiguration.builder().borderColor(borderColor).build());
+    }
+
+    public static Box box(String borderColor) {
+        return new Box(BoxType.ROUNDED, BoxConfiguration.builder().borderColor(borderColor).build());
+    }
+
+    public static Box box(BoxType type, String borderColor) {
+        return new Box(type, BoxConfiguration.builder().borderColor(borderColor).build());
     }
 
 
     // INDENTER
     public static Indenter indenter() {
-        return new DefaultIndenter();
+        return new Indenter();
     }
 
     public static Indenter indenter(IndenterConfiguration indenterConfiguration) {
-        return new DefaultIndenter(indenterConfiguration);
+        return new Indenter(indenterConfiguration);
     }
 
     // PROGRESS BAR
@@ -120,29 +130,35 @@ final class CliqueComponents {
 
     // FRAME
     public static Frame frame() {
-        return new DefaultFrame();
+        return new Frame();
     }
 
     public static Frame frame(FrameConfiguration configuration) {
-        return new DefaultFrame(configuration);
+        return new Frame(configuration);
     }
 
     public static Frame frame(BoxType type) {
-        return new DefaultFrame(type);
+        return new Frame(type);
     }
 
     public static Frame frame(BoxType type, FrameConfiguration configuration) {
-        return new DefaultFrame(configuration, type);
+        return new Frame(configuration, type);
     }
 
-    public static Frame frame(BorderSpec style) {
-        var borderStyle = BorderStyle.fromSpec(style);
-        return new DefaultFrame(FrameConfiguration.fromBorderStyle(borderStyle));
+    public static Frame frame(String borderColor) {
+        return new Frame(FrameConfiguration.builder().borderColor(borderColor).build());
     }
 
-    public static Frame frame(BoxType type, BorderSpec style) {
-        var borderStyle = BorderStyle.fromSpec(style);
-        return new DefaultFrame(FrameConfiguration.fromBorderStyle(borderStyle), type);
+    public static Frame frame(BoxType type, String borderColor) {
+        return new Frame(FrameConfiguration.builder().borderColor(borderColor).build(), type);
+    }
+
+    public static Frame frame(AnsiCode... borderColor) {
+        return new Frame(FrameConfiguration.builder().borderColor(borderColor).build());
+    }
+
+    public static Frame frame(BoxType type, AnsiCode... borderColor) {
+        return new Frame(FrameConfiguration.builder().borderColor(borderColor).build(), type);
     }
