+++ clique-core/src/main/java/io/github/kusoroadeolu/clique/config/FrameConfiguration.java
@@ -1,18 +1,17 @@
 package io.github.kusoroadeolu.clique.config;
 
-
 import io.github.kusoroadeolu.clique.Clique;
-import io.github.kusoroadeolu.clique.core.documentation.InternalApi;
 import io.github.kusoroadeolu.clique.core.documentation.Stable;
+import io.github.kusoroadeolu.clique.core.parser.ParserUtils;
 import io.github.kusoroadeolu.clique.parser.AnsiStringParser;
+import io.github.kusoroadeolu.clique.spi.AnsiCode;
 
+import java.util.Arrays;
 import java.util.Objects;
-
-import static io.github.kusoroadeolu.clique.core.utils.MiscUtils.assertStyleNotNull;
 
 /**
  * @since 3.1.0
- * */
+ */
 @Stable(since = "3.2.0")
 public class FrameConfiguration {
     public static final FrameConfiguration DEFAULT = new FrameConfiguration();
@@ -20,7 +19,7 @@ public class FrameConfiguration {
     private final int padding;
     private final FrameAlign align;
     private final AnsiStringParser parser;
-    private final BorderStyle borderStyle;
+    private final AnsiCode[] borderColor;
 
     private FrameConfiguration() {
         this(new FrameConfigurationBuilder());
@@ -30,7 +29,7 @@ public class FrameConfiguration {
         this.padding = builder.padding;
         this.align = builder.frameAlign;
         this.parser = builder.parser;
-        this.borderStyle = builder.borderStyle;
+        this.borderColor = builder.borderColor;
     }
 
     public static FrameConfigurationBuilder builder() {
@@ -39,24 +38,18 @@ public class FrameConfiguration {
 
     /**
      * @deprecated As of 3.1.3, use {@link FrameConfiguration#builder()} instead. This will be removed in a future release.
-     * */
+     */
     @Deprecated(since = "3.1.3", forRemoval = true)
     public static FrameConfigurationBuilder immutableBuilder() {
         return builder();
-    }
-
-    @InternalApi(since = "3.1.3")
-    public static FrameConfiguration fromBorderStyle(BorderSpec style) {
-        assertStyleNotNull(style);
-        return FrameConfiguration.builder().borderStyle(style).build();
     }
 
     public int getPadding() {
         return this.padding;
     }
 
-    public BorderStyle getBorderStyle() {
-        return this.borderStyle;
+    public AnsiCode[] getBorderColor() {
+        return this.borderColor;
     }
 
     public FrameAlign getFrameAlign() {
@@ -67,12 +60,12 @@ public class FrameConfiguration {
         return this.parser;
     }
 
+    @Override
     public String toString() {
         return "FrameConfiguration[" +
-                "centerPadding=" + padding +
+                "padding=" + padding +
                 ", frameAlign=" + align +
                 ", parser=" + parser +
-                ", borderStyle=" + borderStyle +
                 ']';
     }
 
@@ -81,19 +74,19 @@ public class FrameConfiguration {
         if (object == null || getClass() != object.getClass()) return false;
 
         FrameConfiguration that = (FrameConfiguration) object;
-        return padding == that.padding && align == that.align && parser.equals(that.parser) && Objects.equals(borderStyle, that.borderStyle);
+        return padding == that.padding && align == that.align && parser.equals(that.parser);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(padding, align, parser, borderStyle);
+        return Objects.hash(padding, align, parser, Arrays.hashCode(borderColor));
     }
 
     public static class FrameConfigurationBuilder {
         private int padding = 2;
         private FrameAlign frameAlign = FrameAlign.CENTER;
         private AnsiStringParser parser = Clique.parser();
-        private BorderStyle borderStyle = null;
+        private AnsiCode[] borderColor = new AnsiCode[0];
 
         public FrameConfigurationBuilder padding(int padding) {
             if (padding < 0) throw new IllegalArgumentException("Padding cannot be negative");
@@ -101,14 +94,18 @@ public class FrameConfiguration {
             return this;
         }
 
-        public FrameConfigurationBuilder borderStyle(BorderSpec spec) {
-            Objects.requireNonNull(spec, "Border style cannot be null");
-            this.borderStyle = BorderStyle.fromSpec(spec);
+        public FrameConfigurationBuilder borderColor(String borderColor) {
+            return borderColor(ParserUtils.getAnsiCodes(borderColor, parser).toArray(AnsiCode[]::new));
+        }
+
+        public FrameConfigurationBuilder borderColor(AnsiCode... borderColor) {
+            Objects.requireNonNull(borderColor, "Border color cannot be null");
+            this.borderColor = borderColor;
             return this;
         }
 
         public FrameConfigurationBuilder frameAlign(FrameAlign align) {
-            Objects.requireNonNull(align, "DefaultFrame align cannot be null");
+            Objects.requireNonNull(align, "Frame align cannot be null");
             this.frameAlign = align;
             return this;
         }
