+++ clique-core/src/main/java/io/github/kusoroadeolu/clique/config/TreeConfiguration.java
@@ -1,11 +1,12 @@
 package io.github.kusoroadeolu.clique.config;
 
 import io.github.kusoroadeolu.clique.core.documentation.Stable;
+import io.github.kusoroadeolu.clique.core.parser.ParserUtils;
 import io.github.kusoroadeolu.clique.parser.AnsiStringParser;
+import io.github.kusoroadeolu.clique.spi.AnsiCode;
 
+import java.util.Arrays;
 import java.util.Objects;
-
-import static io.github.kusoroadeolu.clique.core.utils.Constants.EMPTY;
 
 /**
  * @since 3.1.0
@@ -13,7 +14,7 @@ import static io.github.kusoroadeolu.clique.core.utils.Constants.EMPTY;
 @Stable(since = "3.2.0")
 public class TreeConfiguration {
     private final AnsiStringParser parser;
-    private final String guideStyle;
+    private final AnsiCode[] connectorColor;
     public static final TreeConfiguration DEFAULT = new TreeConfiguration();
 
     private TreeConfiguration() {
@@ -22,7 +23,7 @@ public class TreeConfiguration {
 
     private TreeConfiguration(TreeConfigurationBuilder builder) {
         this.parser = builder.parser;
-        this.guideStyle = builder.guideStyle;
+        this.connectorColor = builder.connectorColor;
     }
 
     public static TreeConfigurationBuilder builder() {
@@ -41,8 +42,8 @@ public class TreeConfiguration {
         return parser;
     }
 
-    public String getGuideStyle() {
-        return guideStyle;
+    public AnsiCode[] getConnectorColor() {
+        return connectorColor;
     }
 
     @Override
@@ -50,13 +51,13 @@ public class TreeConfiguration {
         if (object == null || getClass() != object.getClass()) return false;
 
         TreeConfiguration that = (TreeConfiguration) object;
-        return Objects.equals(parser, that.parser) && Objects.equals(guideStyle, that.guideStyle);
+        return Objects.equals(parser, that.parser) && Arrays.equals(connectorColor, that.connectorColor);
     }
 
     @Override
     public int hashCode() {
         int result = Objects.hashCode(parser);
-        result = 31 * result + Objects.hashCode(guideStyle);
+        result = 31 * result + Objects.hashCode(connectorColor);
         return result;
     }
 
@@ -64,12 +65,12 @@ public class TreeConfiguration {
     public String toString() {
         return "TreeConfiguration[" +
                 "parser=" + parser +
-                ", guideStyle='" + guideStyle + '\'' +
+                ", guideStyle='" + Arrays.toString(connectorColor) + '\'' +
                 ']';
     }
 
     public static class TreeConfigurationBuilder {
-        private String guideStyle = EMPTY;
+        private AnsiCode[] connectorColor = new AnsiCode[0];
         private AnsiStringParser parser = AnsiStringParser.DEFAULT;
 
 
@@ -78,12 +79,15 @@ public class TreeConfiguration {
             return this;
         }
 
-        public TreeConfigurationBuilder guideStyle(String guideStyle){
-            Objects.requireNonNull(guideStyle, "Guide style cannot be null");
-            this.guideStyle = guideStyle;
+        public TreeConfigurationBuilder connectorColor(AnsiCode... connectorColor){
+            Objects.requireNonNull(connectorColor, "Connector color cannot be null");
+            this.connectorColor = connectorColor;
             return this;
         }
 
+        public TreeConfigurationBuilder connectorColor(String connectorColor){
+            return connectorColor(ParserUtils.getAnsiCodes(connectorColor, parser).toArray(AnsiCode[]::new));
+        }
 
         public TreeConfiguration build() {
             return new TreeConfiguration(this);
