--- clique-core/src/main/java/io/github/kusoroadeolu/clique/frame/DefaultFrame.java
@@ -1,309 +1,0 @@
-package io.github.kusoroadeolu.clique.frame;
-
-import io.github.kusoroadeolu.clique.boxes.BoxType;
-import io.github.kusoroadeolu.clique.config.BorderStyle;
-import io.github.kusoroadeolu.clique.config.FrameAlign;
-import io.github.kusoroadeolu.clique.config.FrameConfiguration;
-import io.github.kusoroadeolu.clique.core.display.Component;
-import io.github.kusoroadeolu.clique.core.documentation.InternalApi;
-import io.github.kusoroadeolu.clique.core.exceptions.InvalidDimensionException;
-import io.github.kusoroadeolu.clique.core.structures.BorderChars;
-import io.github.kusoroadeolu.clique.core.structures.Cell;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Objects;
-
-import static io.github.kusoroadeolu.clique.ansi.StyleCode.RESET;
-import static io.github.kusoroadeolu.clique.core.utils.BoxUtils.applyAnsiToBorders;
-import static io.github.kusoroadeolu.clique.core.utils.Constants.*;
-import static io.github.kusoroadeolu.clique.core.utils.StringUtils.parseToCellIfPresent;
-
-@InternalApi(since = "3.2.0")
-public class DefaultFrame implements  Frame {
-    private final List<FrameNode> nodes;
-    private final FrameConfiguration configuration;
-    private final BoxType type;
-    private String title;
-    private FrameAlign titleAlign;
-    private int width;
-    private final BorderChars borderChars;
-    private static final int NO_WIDTH_SET = 0;
-    private static final String NULL_FRAME_ALIGN = "DefaultFrame align cannot be null";
-
-    public DefaultFrame(FrameConfiguration configuration, BoxType type) {
-        Objects.requireNonNull(configuration, "Configuration cannot be null");
-        Objects.requireNonNull(type, "DefaultFrame type cannot be null");
-        this.nodes = new ArrayList<>();
-        this.configuration = configuration;
-        this.type = type;
-        this.title = EMPTY;
-        this.titleAlign = FrameAlign.CENTER;
-        this.width = NO_WIDTH_SET;
-        borderChars = BorderChars.from(type);
-        this.styleBorders(borderChars);
-    }
-
-    public DefaultFrame() {
-        this(FrameConfiguration.DEFAULT, BoxType.ROUNDED);
-    }
-
-    public DefaultFrame(BoxType type) {
-        this(FrameConfiguration.DEFAULT, type);
-    }
-
-    public DefaultFrame(FrameConfiguration configuration) {
-        this(configuration, BoxType.ROUNDED);
-    }
-
-    @Override
-    public Frame title(String title, FrameAlign titleAlign) {
-        Objects.requireNonNull(title, "Title cannot be null");
-        Objects.requireNonNull(titleAlign, NULL_FRAME_ALIGN);
-        this.title = title;
-        this.titleAlign = titleAlign;
-        return this;
-    }
-
-    @Override
-    public Frame title(String title) {
-        return title(title, FrameAlign.CENTER);
-    }
-
-    @Override
-    public Frame width(int width) {
-        if (width <= 0) throw new InvalidDimensionException(
-                "Frame width must be greater than zero, got: %d".formatted(width)
-        );        this.width = width;
-        return this;
-    }
-
-    @Override
-    public Frame nest(String str) {
-        return nest(str, configuration.getFrameAlign());
-    }
-
-    @Override
-    public Frame nest(String str, FrameAlign align) {
-        Objects.requireNonNull(str, "String cannot be null");
-        Objects.requireNonNull(align, NULL_FRAME_ALIGN);
-        if (configuration.getParser() != null) str = str + RESET;
-        nodes.add(new FrameNode.StringNode(str, align, configuration.getParser()));
-        return this;
-    }
-
-    @Override
-    public Frame nest(Component component) {
-        return nest(component, configuration.getFrameAlign());
-    }
-
-    @Override
-    public Frame nest(Component component, FrameAlign align) {
-        Objects.requireNonNull(component, "Component cannot be null");
-        Objects.requireNonNull(align, NULL_FRAME_ALIGN);
-        assertNotSelf(component);
-        nodes.add(new FrameNode.ComponentNode(component, align));
-        return this;
-    }
-
-    void assertNotSelf(Component component){
-        if (component == this) throw new IllegalArgumentException("Cannot nest a Frame in itself");
-    }
-
-    public String get() {
-        var appendedTitle = title;
-
-        //If parser is not null
-        if (configuration.getParser() != null && !title.isEmpty()) appendedTitle = title + RESET; //Add a reset flag to prevent title colors from bleeding
-
-
-        var parsedTitle = parseToCellIfPresent(appendedTitle, configuration.getParser());
-
-        //Note that we align the title width by +1 or -1 based on if the frame align is left or right, so to prevent an issue where the title width is left and the frame size = title size, we add one to the width to make up for the by one offset
-        int titleWidth = parsedTitle.width() + 2;
-        int nodesMaxWidth = findNodesMaxWidth(); //Max content width
-
-        int givenWidth = (noWidthSet() ? nodesMaxWidth + (configuration.getPadding() * 2) : this.width);
-
-        if (noWidthSet() && !parsedTitle.isBlank()) {
-            givenWidth = Math.max(givenWidth, titleWidth + 1);
-        }
-
-        int availableWidth = givenWidth - (configuration.getPadding() * 2);
-
-        if (!parsedTitle.isBlank()) validateTitleWidth(titleWidth, givenWidth);
-
-        if (nodesMaxWidth > availableWidth) {
-            throw new InvalidDimensionException(
-                    "Content width (%d) exceeds available frame width (%d). Either increase frame width to at least %d or reduce content size."
-                            .formatted(nodesMaxWidth, availableWidth, nodesMaxWidth + (configuration.getPadding() * 2))
-            );
-        }
-
-        var sb = new StringBuilder();
-        appendTitleToBox(parsedTitle, givenWidth, titleWidth, sb); //Using given width, not available width here, since avail width is meant for content not title
-
-        for (FrameNode node : nodes) {
-            align(node, availableWidth, sb);
-        }
-
-        sb.append(borderChars.bottomLeft())
-                .append(borderChars.hLine().repeat(givenWidth))
-                .append(borderChars.bottomRight())
-                .append(NEWLINE);
-
-        return sb.toString();
-    }
-
-    //Given width = well the given width, avail width = width - (padding * 2), in the case of no width set, the avail width = max node width, given width = (width) + (padding * 2)
-    void align(FrameNode node ,int availableWidth, StringBuilder sb) {
-        var borderChar = this.borderChars;
-        var padding = configuration.getPadding();
-        int rem = Math.max(ZERO, availableWidth - node.maxWidth());
-
-        String fixed = BLANK.repeat(padding);
-        for (var line : node.lines()) {
-            int lineWidth = line.width();
-            String content = line.styledText();
-            switch (node.align()){
-                case RIGHT -> sb.append(borderChar.vLine())
-                        .append(fixed)
-                        .repeat(BLANK,  rem) //We append the remaining width to align right
-                        .append(content)
-                        .append(fixed)
-                        .repeat(BLANK, Math.max(ZERO, availableWidth - lineWidth - rem))
-                        .append(borderChar.vLine())
-                        .append(NEWLINE);
-
-                case LEFT -> {
-                    sb.append(borderChar.vLine())
-                        .append(fixed)
-                        .append(content)
-                        .repeat(BLANK, Math.max(ZERO, availableWidth - lineWidth)) //Just append the remaining space here, we don't need rem here
-                        .append(fixed)
-                        .append(borderChar.vLine())
-                        .append(NEWLINE);
-                }
-
-                case CENTER -> {
-                    int leftPad = rem / 2;
-                    int rightPad = (availableWidth - lineWidth) - leftPad;
-
-                    sb.append(borderChar.vLine())
-                            .append(fixed)
-                            .repeat(BLANK, leftPad) //We need to align everything at a common place
-                            .append(content)
-                            .repeat(BLANK, rightPad)
-                            .append(fixed)
-                            .append(borderChar.vLine())
-                            .append(NEWLINE);
-                }
-            }
-        }
-    }
-
-    void appendTitleToBox(Cell parsedTitle, int givenWidth, int titleWidth, StringBuilder sb){
-        if (!parsedTitle.isBlank()) {
-            int leftWidth = findTitleBlockOffset(givenWidth, titleWidth, titleAlign);
-            sb.append(borderChars.topLeft())
-                    .append(borderChars.hLine().repeat(leftWidth))
-                    .append(BLANK).append(parsedTitle.styledText()).append(BLANK)
-                    //20 - 20 - 1, Gives - 1 For left width
-                    //20 - 20
-
-                    .append(borderChars.hLine().repeat(givenWidth - titleWidth - leftWidth))
-                    .append(borderChars.topRight())
-                    .append(NEWLINE);
-        } else {
-            sb.append(borderChars.topLeft())
-                    .append(borderChars.hLine().repeat(givenWidth))
-                    .append(borderChars.topRight())
-                    .append(NEWLINE);
-        }
-    }
-
-    int findNodesMaxWidth() {
-        return nodes.stream()
-                .mapToInt(FrameNode::maxWidth)
-                .max()
-                .orElse(ZERO);
-    }
-
-    static int findTitleBlockOffset(int givenWidth, int titleWidth, FrameAlign align) {
-        return switch (align) {
-            case LEFT -> 1;
-            case RIGHT -> (givenWidth - titleWidth) - 1; //20 - 20 - 1, Gives  -1, By adding 1 to the title width, we get
-            case CENTER -> (givenWidth - titleWidth) / 2;
-        };
-    }
-
-    boolean noWidthSet(){
-        return this.width == NO_WIDTH_SET;
-    }
-
-    void validateTitleWidth(int titleWidth, int resolvedWidth) {
-        if (titleWidth > resolvedWidth)
-            throw new InvalidDimensionException(
-                    "Title width (%d) exceeds frame width (%d). Increase frame width to at least %d or shorten the title."
-                            .formatted(titleWidth, resolvedWidth, titleWidth)
-            );
-    }
-
-
-    void customizeCorner(char corner) {
-        var str = String.valueOf(corner);
-        if (!str.isBlank()){
-            borderChars.setCorners(str);
-        }
-    }
-
-    void customizeVLine(char vLine) {
-        var str = Character.toString(vLine);
-        if (!str.isBlank()){
-            borderChars.setVLine(str);
-        }
-    }
-
-    void customizeHLine(char hLine) {
-        var str = Character.toString(hLine);
-        if (!str.isBlank()){
-            borderChars.setHLine(str);
-        }
-    }
-
-    void styleBorders(BorderChars borderChar) {
-        final BorderStyle borderStyle = this.configuration.getBorderStyle();
-        if (borderStyle != null) {
-            applyAnsiToBorders(borderChar, borderStyle);
-            customizeHLine(borderStyle.getHorizontalChar());
-            customizeVLine(borderStyle.getVerticalChar());
-            customizeCorner(borderStyle.getCornerChar());
-        }
-    }
-
-    @Override
-    public boolean equals(Object object) {
-        if (object == null || getClass() != object.getClass()) return false;
-
-        DefaultFrame frame = (DefaultFrame) object;
-        return width == frame.width && Objects.equals(nodes, frame.nodes) && Objects.equals(configuration, frame.configuration) && type == frame.type && Objects.equals(title, frame.title) && titleAlign == frame.titleAlign && Objects.equals(borderChars, frame.borderChars);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(width, nodes, configuration, type, title, titleAlign, borderChars);
-    }
-
-    @Override
-    public String toString() {
-        return "Frame[" +
-                "borderChars=" + borderChars +
-                ", width=" + width +
-                ", titleAlign=" + titleAlign +
-                ", title='" + title + '\'' +
-                ", type=" + type +
-                ", configuration=" + configuration +
-                ", nodes=" + nodes +
-                ']';
-    }
-}
