+++ clique-core/src/main/java/io/github/kusoroadeolu/clique/tables/DefaultTable.java
@@ -1,18 +1,16 @@
 package io.github.kusoroadeolu.clique.tables;
 
 
-import io.github.kusoroadeolu.clique.config.BorderStyle;
 import io.github.kusoroadeolu.clique.config.TableConfiguration;
 import io.github.kusoroadeolu.clique.core.structures.WidthAwareList;
 import io.github.kusoroadeolu.clique.core.utils.Constants;
-import io.github.kusoroadeolu.clique.style.DefaultStyleBuilder;
-import io.github.kusoroadeolu.clique.style.StyleBuilder;
 
 import java.util.Objects;
 
 import static io.github.kusoroadeolu.clique.core.utils.StringUtils.clearStringBuilder;
 import static io.github.kusoroadeolu.clique.core.utils.TableUtils.align;
 import static io.github.kusoroadeolu.clique.core.utils.TableUtils.chooseColAlignment;
+import static io.github.kusoroadeolu.clique.style.StyleBuilder.formatAndReset;
 
 class DefaultTable extends AbstractTable {
     private String corner;
@@ -24,7 +22,7 @@ class DefaultTable extends AbstractTable {
         this.corner = "+";
         this.hLine = "-";
         this.vLine = "|";
-        this.styleTableBorders();
+        this.colorTableBorders();
 
     }
 
@@ -35,7 +33,7 @@ class DefaultTable extends AbstractTable {
         final var tableBuilder = new StringBuilder();
         final var sb = new StringBuilder();
         final var headerAndFooter = this.appendHeader(sb);
-        final int padding = this.tableConfiguration.getPadding();
+        final int padding = this.configuration.getPadding();
         clearStringBuilder(sb);
 
 
@@ -46,7 +44,7 @@ class DefaultTable extends AbstractTable {
             tableBuilder.append(vLine);
 
             for (int j = 0; j < list.size(); j++) {
-                var cellAlign = this.tableConfiguration.getAlignment();
+                var cellAlign = this.configuration.getAlignment();
                 final String styledCell = list.getStyledText(j);
                 final int displayWidth = list.get(j).width();
                 final WidthAwareList cl = this.columns.get(j);
@@ -54,7 +52,7 @@ class DefaultTable extends AbstractTable {
 
                 final int offset = (longest - displayWidth) + padding;
 
-                cellAlign = chooseColAlignment(j, cellAlign, this.tableConfiguration.getColumnAlignment());
+                cellAlign = chooseColAlignment(j, cellAlign, this.configuration.getColumnAlignment());
                 tableBuilder.append(align(cellAlign, sb, offset, styledCell, vLine));
                 clearStringBuilder(sb);
             }
@@ -70,48 +68,20 @@ class DefaultTable extends AbstractTable {
     private String appendHeader(StringBuilder sb) {
         for (final WidthAwareList l : this.columns) {
             sb.append(corner);
-            sb.repeat(hLine, l.longest() + this.tableConfiguration.getPadding());
+            sb.repeat(hLine, l.longest() + this.configuration.getPadding());
         }
         sb.append(corner);
         return sb.toString();
     }
 
-    public void customizeCorner(char corner) {
-        var str = Character.toString(corner);
-        if (!str.isBlank()){
-            this.corner = str;
-            nullCachedString();
-        }
 
-    }
-
-    public void customizeVLine(char vLine) {
-        var str = Character.toString(vLine);
-        if (!str.isBlank()){
-            this.vLine = str;
-            nullCachedString();
-        }
-    }
-
-    public void customizeHLine(char hLine) {
-        var str = Character.toString(hLine);
-        if (!str.isBlank()){
-            this.hLine = str;
-            nullCachedString();
-        }
-    }
-
-
-    protected void styleTableBorders() {
-        final BorderStyle borderStyle = this.tableConfiguration.getBorderStyle();
-        if (borderStyle != null){
-            final StyleBuilder sb = new DefaultStyleBuilder();
-            this.hLine = sb.formatAndReset(this.hLine, borderStyle.getHorizontalStyle());
-            this.vLine = sb.formatAndReset(this.vLine, borderStyle.getVerticalStyle());
-            this.corner = sb.formatAndReset(this.corner, borderStyle.getCornerStyle());
-            this.customizeCorner(borderStyle.getCornerChar());
-            this.customizeHLine(borderStyle.getHorizontalChar());
-            this.customizeVLine(borderStyle.getVerticalChar());
+    protected void colorTableBorders() {
+        var color = configuration.getBorderColor();
+        if (color != null && color.length != 0){
+            final var sb = new StringBuilder();
+            this.hLine = formatAndReset(sb, this.hLine, color);
+            this.vLine = formatAndReset(sb, this.vLine, color);
+            this.corner = formatAndReset(sb, this.corner, color);
         }
 
     }
@@ -133,7 +103,7 @@ class DefaultTable extends AbstractTable {
     @Override
     public String toString() {
         return "DefaultTable[" +
-                "tableConfiguration=" + tableConfiguration +
+                "tableConfiguration=" + configuration +
                 ", vLine='" + vLine + '\'' +
                 ", hLine='" + hLine + '\'' +
                 ", edge='" + corner + '\'' +
