+++ clique-core/src/main/java/io/github/kusoroadeolu/clique/tree/Tree.java
@@ -2,10 +2,8 @@ package io.github.kusoroadeolu.clique.tree;
 
 import io.github.kusoroadeolu.clique.config.TreeConfiguration;
 import io.github.kusoroadeolu.clique.core.display.Borderless;
-import io.github.kusoroadeolu.clique.core.parser.ParserUtils;
 import io.github.kusoroadeolu.clique.core.utils.StringUtils;
 import io.github.kusoroadeolu.clique.spi.AnsiCode;
-import io.github.kusoroadeolu.clique.style.DefaultStyleBuilder;
 import io.github.kusoroadeolu.clique.style.StyleBuilder;
 
 import java.util.ArrayList;
@@ -23,7 +21,7 @@ public class Tree implements Borderless {
     static final String END_CONNECTOR = "└─ ";
     static final String SPACE = "   ";
     static final String CONNECTING_LINE = "│  ";
-    private final AnsiCode[] guideStyle;
+    private final AnsiCode[] connectorColor;
     private Tree parent;
 
     public Tree(String label) {
@@ -41,7 +39,7 @@ public class Tree implements Borderless {
         this.children = new ArrayList<>();
         this.treeConfiguration = treeConfiguration;
         this.parent = parent;
-        this.guideStyle = ParserUtils.getAnsiCodes(treeConfiguration.getGuideStyle()).toArray(AnsiCode[]::new);
+        this.connectorColor = treeConfiguration.getConnectorColor();
     }
 
     public Tree add(String label) {
@@ -65,7 +63,7 @@ public class Tree implements Borderless {
 
     private void buildTree(Tree node, String prefix, boolean isLast, StyleBuilder sb) {
         String connector = isLast ? END_CONNECTOR : CONNECTOR;
-        sb.stack(prefix, this.guideStyle)
+        sb.stack(prefix, this.connectorColor)
                 .append(connector)
                 .append(node.label)
                 .append(NEWLINE);
@@ -90,7 +88,7 @@ public class Tree implements Borderless {
 
     @Override
     public String get() {
-        var sb = new DefaultStyleBuilder();
+        var sb = new StyleBuilder();
         sb.append(label).append(NEWLINE);
         for (int i = 0; i < children.size(); i++) {
             buildTree(children.get(i), EMPTY, i == children.size() - 1, sb);
